+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - **BREAKING** feat: Rename `makeVariables` helper (remove `variables`)
+- feat: Add `makeVariables.isVar()`
 
 ## 0.2.10